@@ -1,115 +1,77 @@
 {
   "name": "almanime-frontend",
-  "version": "0.1.0",
   "private": true,
-  "engines": {
-    "node": "20.x"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
     "postinstall": "node ./copy-dep.js",
-    "start": "BROWSER=none react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --coverage",
-    "test:all": "react-scripts test --coverage --watchAll",
-    "lint": "eslint src",
-    "eject": "react-scripts eject",
-    "storybook": "start-storybook -p 6006 -s public",
-    "build-storybook": "build-storybook -s public"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "test": "vitest run --coverage",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0 && stylelint '**/*.scss'",
+    "prettier:check": "prettier --check src/",
+    "prettier:write": "prettier --write src/",
+    "preview": "vite preview",
+    "storybook": "storybook dev -p 6006",
+    "build-storybook": "storybook build"
   },
   "dependencies": {
     "@almanime/srt-to-ass": "^1.0.7",
-    "@auth0/auth0-react": "^2.0.0",
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11",
-    "@mdi/js": "^7.1.96",
+    "@auth0/auth0-react": "^2.2.4",
+    "@mdi/js": "^7.4.47",
     "@mdi/react": "^1.6.1",
-    "@sentry/react": "^7.36.0",
-    "@sentry/tracing": "^7.36.0",
-    "@testing-library/jest-dom": "^6.1.2",
+    "@sentry/react": "^8.5.0",
     "@videojs/themes": "^1.0.1",
-    "ass-compiler": "^0.1.7",
+    "ass-compiler": "^0.1.11",
     "bulma": "^0.9.4",
-    "history": "^5.3.0",
-    "jest-fetch-mock": "^3.0.3",
     "libass-wasm": "4.0.0",
-    "luxon": "^3.2.1",
-    "path-to-regexp": "^6.2.1",
+    "luxon": "^3.4.4",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-dropzone": "^14.2.3",
     "react-query": "^3.39.3",
-    "react-router-dom": "^6.8.0",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.23.1",
     "react-user-avatar": "^1.10.0",
-    "sass": "^1.58.0",
-    "typescript": "^5.2.2",
-    "video.js": "^8.0.3",
-    "wavesurfer.js": "^7.2.2",
-    "web-vitals": "^3.1.1"
+    "sass": "^1.77.2",
+    "video.js": "8.12.0",
+    "wavesurfer.js": "^7.7.15",
+    "web-vitals": "^4.0.1"
   },
   "devDependencies": {
-    "@storybook/addon-actions": "^7.4.0",
-    "@storybook/addon-essentials": "^7.4.0",
-    "@storybook/addon-links": "^7.4.0",
-    "@storybook/builder-webpack5": "^7.4.0",
-    "@storybook/manager-webpack5": "^6.5.16",
-    "@storybook/node-logger": "^7.4.0",
-    "@storybook/preset-create-react-app": "^7.4.0",
-    "@storybook/react": "^7.4.0",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.4.3",
-    "@types/jest": "^29.4.0",
-    "@types/luxon": "^3.2.0",
-    "@types/node": "^20.5.9",
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
-    "@types/react-router-dom": "^5.3.3",
-    "@typescript-eslint/eslint-plugin": "^6.5.0",
-    "@typescript-eslint/parser": "^6.5.0",
-    "eslint": "^8.33.0",
-    "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-react-app": "^7.0.1",
-    "eslint-plugin-import": "^2.27.5",
-    "eslint-plugin-jest": "^27.2.1",
-    "eslint-plugin-jsx-a11y": "^6.7.1",
-    "eslint-plugin-react": "^7.32.2",
+    "@chromatic-com/storybook": "^1.5.0",
+    "@storybook/addon-essentials": "^8.1.5",
+    "@storybook/addon-interactions": "^8.1.5",
+    "@storybook/addon-links": "^8.1.5",
+    "@storybook/addon-onboarding": "^8.1.5",
+    "@storybook/blocks": "^8.1.5",
+    "@storybook/react": "^8.1.5",
+    "@storybook/react-vite": "^8.1.5",
+    "@storybook/test": "^8.1.5",
+    "@testing-library/react": "^15.0.7",
+    "@types/luxon": "^3",
+    "@types/react": "^18.2.66",
+    "@types/react-dom": "^18.2.22",
+    "@typescript-eslint/eslint-plugin": "^7.2.0",
+    "@typescript-eslint/parser": "^7.2.0",
+    "@vitejs/plugin-basic-ssl": "^1.1.0",
+    "@vitejs/plugin-react-swc": "^3.5.0",
+    "@vitest/coverage-v8": "^1.6.0",
+    "eslint": "^8.57.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-simple-import-sort": "^10.0.0",
-    "eslint-plugin-storybook": "^0.6.10",
-    "storybook-dark-mode": "^3.0.1",
-    "stylelint": "^15.10.3",
-    "stylelint-config-sass-guidelines": "^10.0.0",
-    "webpack": "^5.75.0"
+    "eslint-plugin-react-refresh": "^0.4.6",
+    "eslint-plugin-storybook": "^0.8.0",
+    "jsdom": "^24.1.0",
+    "prettier": "3.2.5",
+    "storybook": "^8.1.5",
+    "stylelint": "^16.6.1",
+    "stylelint-config-sass-guidelines": "^11.1.0",
+    "typescript": "^5.2.2",
+    "vite": "^5.2.0",
+    "vitest": "^1.6.0"
   },
-  "browserslist": {
-    "production": [
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version",
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
-    ]
-  },
-  "jest": {
-    "transformIgnorePatterns": [
-      "node_modules/(?!wavesurfer.js)/"
-    ],
-    "coverageThreshold": {
-      "global": {
-        "statements": 23.18,
-        "branches": 11.89,
-        "functions": 15.44,
-        "lines": 23.11
-      }
-    }
+  "resolutions": {
+    "tough-cookie": "5.0.0-rc.2"
   },
   "packageManager": "yarn@4.2.2"
 }
