@@ -1,18 +1,11 @@
-import { createRoot } from 'react-dom/client';
-import { act } from '@testing-library/react';
-
+import { render } from '@testing-library/react';
+import { test } from 'vitest';
 import { Error, Success } from './help';
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  const root = createRoot(div);
-
-  act(() => root.render(<Error message="Test" />));
+test('renders without crashing', () => {
+  render(<Error message="Test" />);
 });
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  const root = createRoot(div);
-
-  act(() => root.render(<Success message="Test" />));
+test('renders without crashing', () => {
+  render(<Success message="Test" />);
 });
