@@ -1,13 +1,9 @@
-import { createRoot } from 'react-dom/client';
-import { act } from '@testing-library/react';
-
+import { render } from '@testing-library/react';
+import { test } from 'vitest';
 import Pagination from './pagination';
 
-it('renders without crashing', () => {
-  const div = document.createElement('div');
-  const root = createRoot(div);
-
-  act(() => root.render(
+test('renders without crashing', () => {
+  render(
     <Pagination
       total={10}
       perPage={2}
@@ -15,5 +11,5 @@ it('renders without crashing', () => {
       steps={3}
       onChange={() => {}}
     />,
-  ));
+  );
 });
