@@ -1,19 +1,19 @@
-import { ReportHandler } from 'web-vitals';
+import { Metric } from 'web-vitals';
 
-const reportWebVitals = (onPerfEntry?: ReportHandler): void => {
+const reportWebVitals = (onPerfEntry?: (metric: Metric) => void): void => {
   if (onPerfEntry && onPerfEntry instanceof Function) {
     import('web-vitals').then(({
-      getCLS,
-      getFID,
-      getFCP,
-      getLCP,
-      getTTFB,
+      onCLS,
+      onINP,
+      onLCP,
+      onFCP,
+      onTTFB,
     }) => {
-      getCLS(onPerfEntry);
-      getFID(onPerfEntry);
-      getFCP(onPerfEntry);
-      getLCP(onPerfEntry);
-      getTTFB(onPerfEntry);
+      onCLS(onPerfEntry);
+      onINP(onPerfEntry);
+      onLCP(onPerfEntry);
+      onFCP(onPerfEntry);
+      onTTFB(onPerfEntry);
     });
   }
 };
