@@ -1,4 +1,4 @@
-import FansubRole from 'enums/FansubRole';
+import FansubRole from "../enums/FansubRole";
 
 export type Member = {
   name: string;
